refactor: move paste code to separate component
